% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\name{daemon}
\alias{daemon}
\title{Start the \code{IsoplotR} GUI without exiting}
\usage{
daemon(port = NULL, host = "127.0.0.1")
}
\arguments{
\item{port}{Internet port of the virtual server. If not defined, a
random free port will be chosen and the browser will be opened
to show the GUI.}

\item{host}{IP address of the virtual server}
}
\value{
This function does not return.
}
\description{
Opens a web-browser with a Graphical User Interface (GUI) for the
\code{IsoplotR} package. This function is intended to be used from
Rscript so that Rscript does not terminate and the server stays up.
}
\examples{
# runs indefinitely unless interrupted by the user:
\dontrun{daemon(3838)}
}
