% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{encodePlot}
\alias{encodePlot}
\title{Renders a plot as a base64-encoded image}
\usage{
encodePlot(device, mimeType, width, height, plotFn)
}
\arguments{
\item{device}{Graphics device function, such as [grDevices::png]
or [grDevices::pdf]}

\item{mimeType}{Mime type for the data produced by `device`}

\item{width}{Width of the plot in units applicable to `device`}

\item{height}{Height of the plot in units applicable to `device`}

\item{plotFn}{Function to call to perform the plot}
}
\description{
Renders a plot as a base64-encoded image
}
\seealso{
[encodePlotAsPng()]

[encodePlotAsPdf()]
}
