% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{encodePlotAsPdf}
\alias{encodePlotAsPdf}
\title{Renders a plot as a base64-encoded PDF}
\usage{
encodePlotAsPdf(width, height, plotFn)
}
\arguments{
\item{width}{Width of the plot in inches}

\item{height}{Height of the plot in inches}

\item{plotFn}{Function to call to perform the plot}
}
\description{
The result can be set as the `href` attribute of an `a` element in HTML
to allow the PDF to be downloaded (also set a `download` attribute to
a reasonable filename).
}
