% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{encodePlotAsPng}
\alias{encodePlotAsPng}
\title{Renders a plot as a base64-encoded PNG}
\usage{
encodePlotAsPng(width, height, plotFn)
}
\arguments{
\item{width}{Width of the plot in pixels}

\item{height}{Height of the plot in pixels}

\item{plotFn}{Function to call to perform the plot}
}
\description{
The result can be set as the `src` attribute of an `img` element in HTML.
}
\seealso{
[rrpcServer()]
}
