% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\name{IsoplotR}
\alias{IsoplotR}
\title{Start the \code{IsoplotR} GUI}
\usage{
IsoplotR(host = "0.0.0.0", port = 8080)
}
\arguments{
\item{host}{IP address of the virtual server}

\item{port}{internet port of the virtual server}
}
\description{
Opens a web-browser with a Graphical User Interface (GUI) for the
\code{IsoplotR} package. An online version of the same interface is
provided at \url{http://isoplotr.london-geochron.com}
}
\examples{
#IsoplotR()
}
