% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\name{stopIsoplotR}
\alias{stopIsoplotR}
\title{Stop an \code{IsoplotR} GUI}
\usage{
stopIsoplotR(server = NULL)
}
\arguments{
\item{server}{The server (returned by [IsoplotR()]) to stop. If not
supplied all servers will be stopped.}
}
\value{
The number of servers stopped.
}
\description{
Stop an \code{IsoplotR} GUI
}
\examples{
#IsoplotR()
}
