% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\name{IsoplotR}
\alias{IsoplotR}
\title{Start the \code{IsoplotR} GUI}
\usage{
IsoplotR()
}
\description{
Opens a web-browser with a Graphical User Interface (GUI) for the
\code{IsoplotR} package. An online version of the same interface is
provided at \url{http://isoplotr.london-geochron.com}
}
\examples{
#IsoplotR()
}

