% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\name{IsoplotR}
\alias{IsoplotR}
\title{Start the \code{IsoplotR} GUI}
\usage{
IsoplotR(
  host = if (daemonize) "0.0.0.0" else "127.0.0.1",
  port = NULL,
  timeout = Inf,
  daemonize = !is.null(port)
)
}
\arguments{
\item{host}{IP address of the virtual server, default is 0.0.0.0}

\item{port}{Internet port of the virtual server. If not defined, a
random free port will be chosen and the browser will be opened
to show the GUI.}

\item{timeout}{How long (in wall-clock time) an operation may take
before returning with a timeout error. Default: no limit.}

\item{daemonize}{logical. If \code{TRUE}, runs as daemon.}
}
\value{
server object
}
\description{
Opens a web-browser with a Graphical User Interface (GUI) for the
\code{IsoplotR} package. An online version of the same interface is
provided at \url{https://www.ucl.ac.uk/~ucfbpve/isoplotr/}
}
\examples{
\donttest{IsoplotR()}
}
