% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoplotR.R
\name{IsoplotR}
\alias{IsoplotR}
\title{Start the \code{IsoplotR} GUI}
\usage{
IsoplotR(host = "0.0.0.0", port = NULL)
}
\arguments{
\item{host}{IP address of the virtual server, default is 0.0.0.0}

\item{port}{Internet port of the virtual server. If not defined, a
random free port will be chosen and the browser will be opened
to show the GUI.}
}
\value{
server object
}
\description{
Opens a web-browser with a Graphical User Interface (GUI) for the
\code{IsoplotR} package. An online version of the same interface is
provided at \url{https://www.ucl.ac.uk/~ucfbpve/isoplotr/}
}
\examples{
\donttest{IsoplotR()}
}
